+++ apps/backend/src/modules/chats/chat.repository.ts
@@ -213,14 +213,22 @@ class ChatRepository implements Repository<ChatEntity> {
 				void builder
 					.where((builder) => {
 						void builder
-							.whereILike("firstUser:userDetails.firstName", `%${search}%`)
-							.orWhereILike("firstUser:userDetails.lastName", `%${search}%`);
+							.whereILike("firstUser:userDetails.lastName", `%${search}%`)
+							.orWhereRaw("concat(??, ' ', ??) ILIKE ?", [
+								"firstUser:userDetails.firstName",
+								"firstUser:userDetails.lastName",
+								`%${search}%`,
+							]);
 					})
 					.andWhere({ secondUserId: userId })
 					.orWhere((builder) => {
 						void builder
-							.whereILike("secondUser:userDetails.firstName", `%${search}%`)
-							.orWhereILike("secondUser:userDetails.lastName", `%${search}%`);
+							.whereILike("secondUser:userDetails.lastName", `%${search}%`)
+							.orWhereRaw("concat(??, ' ', ??) ILIKE ?", [
+								"secondUser:userDetails.firstName",
+								"secondUser:userDetails.lastName",
+								`%${search}%`,
+							]);
 					})
 					.andWhere({ firstUserId: userId });
 			})