+++ apps/frontend/src/pages/chats/libs/components/chat-link/chat-link.tsx
@@ -5,7 +5,9 @@ import {
 	configureString,
 	getDifferenceInHours,
 	getFormattedDate,
+	getValidClassNames,
 } from "~/libs/helpers/helpers.js";
+import { useAppSelector } from "~/libs/hooks/hooks.js";
 import { type ChatGetAllItemResponseDto } from "~/modules/chats/chats.js";
 
 import styles from "./styles.module.css";
@@ -15,6 +17,9 @@ type Properties = {
 };
 
 const ChatLink: React.FC<Properties> = ({ chat }: Properties) => {
+	const { currentChat } = useAppSelector(({ chats }) => ({
+		currentChat: chats.currentChat,
+	}));
 	const { id, interlocutor, lastMessage } = chat;
 	const chatRouteById = configureString(AppRoute.CHATS_$ID, {
 		id: String(id),
@@ -26,8 +31,16 @@ const ChatLink: React.FC<Properties> = ({ chat }: Properties) => {
 			? getFormattedDate(lastMessage.createdAt, FormatDateType.HH_MM)
 			: getFormattedDate(lastMessage.createdAt, FormatDateType.DD_MM_YYYY);
 
+	const isCurrentChat = currentChat?.id === id;
+
 	return (
-		<Link className={styles["container"]} to={chatRouteById}>
+		<Link
+			className={getValidClassNames(
+				styles["container"],
+				isCurrentChat && styles["current"],
+			)}
+			to={chatRouteById}
+		>
 			<Image
 				alt="User avatar"
 				height="48"
@@ -36,15 +49,20 @@ const ChatLink: React.FC<Properties> = ({ chat }: Properties) => {
 				width="48"
 			/>
 			<div className={styles["text-container"]}>
-				<div className={styles["top-container"]}>
+				<div className={styles["message-container"]}>
 					<span
 						className={styles["name"]}
 					>{`${interlocutor.firstName} ${interlocutor.lastName}`}</span>
-					<span className={styles["date"]}>{date}</span>
+
+					<p className={styles["message"]}>
+						{isLastMessagesYours && (
+							<span className={styles["your-message"]}>You: </span>
+						)}
+						{lastMessage.text}
+					</p>
 				</div>
-				<p className={styles["message"]}>
-					{isLastMessagesYours ? `You: ${lastMessage.text}` : lastMessage.text}
-				</p>
+
+				<span className={styles["date"]}>{date}</span>
 			</div>
 		</Link>
 	);